Implement a ping mechanism
Pings are useful to detect whether a worker is working properly or is
stuck. I define "working" to be polling for new jobs to process. It
would be wasteful to ping every time it polls for jobs and the ping
shouldn't disrupt the dequeue command, so execute it on a separate
async task and only send a ping every 5 seconds when there is activity.