Rework the queued job implementation
Prior to this, I was seeing a number of errors upon enqueuing that
looked like:

> duplicate key value violates unique constraint "queued_job_pkey"

These were happening because two transactions were overlapping and both
reading the same `min(position)` value and attempting to enqueue a job
for `position + 1`. This wouldn't happen at the SERIALIZABLE isolation
level, but it would require retrying the transaction.

So that this still works at READ COMMITTED, I came up with a slightly
different approach. `queued_job.position` is now generated as a sequence
so that jobs within a queue are ordered. This is low overhead and there
should be no contention to insert into this.

Then we need to make sure that the jobs are moved to `running_job` in
the correct order. `running_job.queue` ensures that only one job for a
given queue is running at a time. The `enqueue` call attempts to put the
job into `running_job` and gives up if there is already a job being
processed. The worker will pull the next job into `running_job` when
finishing the current one. Thus, `enqueue` should only put a job into
`running_job` when that queue is empty. This seemed to work empirically
when benchmarking this.