Handle the situation where a blocking job has finished
When enqueuing a job blocked on another job, this previously assumed
that the blocking job had not been executed yet. The worker will find
all blocked jobs when finishing a job and move those to running_job, as
part of its bookkeeping. However, if the job has already finished, this
mechanism won't work and the blocked job will never run.

That could be addressed by having an "unblocker" component that checks
rows in blocked_job and moves them to running job if the blocking_job_id
is in finished_job. That adds operational complexity though. I can
achieve the same effect by checking upon insertion.