Yet another approach to handling blocked jobs
This introduces a task that runs periodically to find blocked jobs where
the blocking job has already finished. Despite trying to be careful in
the implementation of both enqueue and worker, consistently a worker
would finish a blocking job and both (A) not see that a new job was
blocked by it and (B) the enqueue not see that the blocking job had
finished. I thought about this for a bit and am not sure this can be
solved in those two transactions at the READ COMMITTED level, but I may
be missing an approach. This seems like a inelegant compromise that
works sufficiently well. It also has the benefit of minimizing the
amount of work that happens upon enqueue and gets some batching with
unblocker's DELETE.