Run multiple workers in the same process
I evaluated different approaches, but this strikes a good balance of
being a simple implementation and being responsive. The `running_job`
table will reflect the true state of job processing, `worker` will have
an entry for each thread that is processing something. Worker tasks will
complete with each other to dequeue tasks and that could be batched, but
SKIP LOCKED should avoid a major performance hit.